backend/requirements.txt
Django==5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-environ
psycopg2-binary
python-dateutil
python-dotenv
python-decouple
icalendar
